move mungegithub prow job utils out of submit-queue-batch.go into separate file
